pyproject.toml
[project]
name = "vrchat-guide"
version = "0.1.0"
description = "Intelligent VRChat Guide using Genie Worksheets framework"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }

# Core dependencies with version constraints
dependencies = [
    "pip==24.3.1",
    # Genie Worksheets Core Dependencies
    "chainlite < 0.3.0",
    "google-api-python-client>=2.149.0",
    "google-auth>=2.35.0",
    "google_auth_oauthlib>=1.2.1",
    "json-repair>=0.30.0",
    "loguru>=0.7.2",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "sympy>=1.13.3",
    "termcolor>=2.5.0",
    "toml>=0.10.2",
    "sql-metadata>=2.13.0",
    "pydantic>=2.0.0",
    "PyYAML>=6.0.1",
    "langchain-openai>=0.1.7",
    "langchain-together>=0.1.1",
    "grandalf>=0.8",
    
    # VRChat Integration Core Dependencies
    "numpy>=1.24.0",
    "sentence-transformers>=2.2.2",
    "python-dotenv>=1.0.0",
    "unidecode>=1.3.6",
    "pymongo>=4.5.0",
    "python-osc"
]

# Optional feature sets
[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "pytest>=7.0.0",
    "pylint>=2.17.0",
    "jupyter>=1.1.1",
]

suql = [
    "Jinja2==3.1.2",
    "Flask==2.3.2",
    "Flask-Cors==4.0.0",
    "Flask-RESTful==0.3.10",
    "requests==2.31.0",
    "tiktoken>=0.4.0",
    "psycopg2-binary>=2.9.7",
    "pglast",
    "litellm>=1.34.34",
    "platformdirs>=4.0.0",
    "faiss-cpu>=1.7.4",  # or faiss-gpu for GPU support
    "spacy>=3.5.0",
    "litellm>=0.1.1",
    "psycopg2-binary>=2.9.9",
    "flagembedding==1.3.2",  
]


web = [
    "chainlit>=1.2.0",
]

voice = [
    "deepgram-sdk==2.12.0",
    "sounddevice>=0.4.6",
    "speechrecognition>=3.10.0",
    "playsound==1.2.2",
    "pydub>=0.25.1",
    "vosk>=0.3.45",
    "pyaudio>=0.2.13",
    "soundfile>=0.12.1",
    "wavio>=0.0.7",
    "ffmpeg-python>=0.2.0",
]

cloud = [
    "boto3>=1.28.0",
]

all = [
    "vrchat-guide[suql,dev,web,voice,cloud]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vrchat_guide"]

[tool.hatch.build]
only-packages = true
artifacts = [
    "src/vrchat_guide/py.typed",
]

[tool.uv.sources]
genie-worksheets = { path = "packages/genie-worksheets", editable = true }
neu-llm-avatars = { path = "packages/neu-llm-avatars", editable = true }

# Optional features
kraken = { path = "packages/genie-worksheets/packages/knowledge-agent", editable = true }
suql = { path = "packages/suql", editable = true }